+++ packages/drizzle-driver/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @powersync/drizzle-driver
 
+## 0.3.2
+
+### Patch Changes
+
+- 6580f29: Added support for custom column types when converting a Drizzle schema to a PowerSync app schema.
+
+## 0.3.1
+
+### Patch Changes
+
+- 86a753f: Fixed Drizzle transactions breaking for react-native projects, correctly using lock context for transactions.
+
 ## 0.3.0
 
 ### Minor Changes